fix(backend): Resolve 404 for /api/documents/upload by refactoring module
- Resolves TASK-012.
- Created `backend/app/modules/__init__.py` to ensure `modules` is a package.
- Refactored the `backend/app/modules/documents/` module to align with the v1 structure used by other modules:
    - Consolidated endpoints into `backend/app/modules/documents/v1/endpoints.py`.
    - Consolidated schemas into `backend/app/modules/documents/v1/schemas.py`.
    - Deleted now-redundant `backend/app/modules/documents/endpoints.py` and `schemas.py`.
- Updated `backend/app/config/modules.yaml` to correctly point to the refactored `documents` module path (`modules.documents.v1`) and router variable name (`router`).
- Updated `TASK-012` in jules-flow to document the fix and mark as done.
